Clear test-target 1-warning clippy batch across 38 crates
Sweeping fixes for warnings that only fire on test builds (no warnings
in --bin path), uncovered by adding --tests to the clippy invocation.

Mechanical patterns applied:
- assert_eq!(x, false) / (x, true)            -> assert!(!x) / assert!(x)
  (apps/backup, init/login, charmap, ...)
- assert!(a >= N && a <= M)                   -> assert!((N..=M).contains(&a))
  (apps/calendar, apps/netscan)
- len() >= 1 / > 0                            -> !is_empty()
  (apps/explorer, apps/charmap)
- map_or(default, ...)                        -> is_some_and / is_none_or
  (apps/launcher, apps/match3)
- needless mut / borrow / unnecessary cast    -> remove
  (apps/partmanager, apps/rssreader, gui/toolkit, capsh)
- field-assign-after-Default                  -> struct-update syntax
  (apps/imageviewer, apps/camera, apps/screenrecorder)
- collapsible if/let chains                   -> let-chain &&
  (apps/solitaire, userspace/awk, sed, jq)
- needless return-of-let-binding              -> tail expression
  (apps/freecell)
- unused std::env in #[cfg(not(test))] main   -> gated import
  (userspace/blkid, ss, acpi)
- unused vars                                 -> underscore prefix
  (userspace/iw, hdparm, column)
- float strict-equality assert                -> abs-diff < EPSILON
  (apps/passwordgen)
- Misc: no-op shift, unnecessary cfg-test get_args, identical if/elif
  collapsed via single boolean, vec![] -> array, octal grouping, etc.

Also: scoped #[allow(too_many_arguments)] on test-only wtmp record
builder in userspace/last (mirrors on-disk record layout); scoped
#[allow(dead_code)] on TestEnv::{read_file, write_file} in
userspace/useradd (reserved for future round-trip tests).

+++ apps/partmanager/src/main.rs
@@ -3947,7 +3947,7 @@ mod tests {
     fn test_create_partition_dialog_selected_size() {
         let mut d = CreatePartitionDialog::new(0, 2048, 512);
         d.size_percent = 50;
-        let expected = (2048u64 * 512 / 2) as u64;
+        let expected = 2048u64 * 512 / 2;
         assert_eq!(d.selected_size_bytes(), expected);
     }
